+++ example/lib/main.dart
@@ -11,18 +11,20 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  static const UUID = '39ED98FF-2900-441A-802F-9C398FC199D2';
-  static const MAJOR_ID = 1;
-  static const MINOR_ID = 100;
-  static const TRANSMISSION_POWER = -59;
-  static const IDENTIFIER = 'com.example.myDeviceRegion';
-  static const LAYOUT = BeaconBroadcast.ALTBEACON_LAYOUT;
-  static const MANUFACTURER_ID = 0x0118;
+  static const String uuid = '39ED98FF-2900-441A-802F-9C398FC199D2';
+  static const int majorId = 1;
+  static const int minorId = 100;
+  static const int transmissionPower = -59;
+  static const String identifier = 'com.example.myDeviceRegion';
+  static const AdvertiseMode advertiseMode = AdvertiseMode.lowPower;
+  static const String layout = BeaconBroadcast.ALTBEACON_LAYOUT;
+  static const int manufacturerId = 0x0118;
+  static const List<int> extraData = [100];
 
   BeaconBroadcast beaconBroadcast = BeaconBroadcast();
 
-  BeaconStatus _isTransmissionSupported;
   bool _isAdvertising = false;
+  BeaconStatus _isTransmissionSupported;
   StreamSubscription<bool> _isAdvertisingSubscription;
 
   @override
@@ -34,8 +36,7 @@ class _MyAppState extends State<MyApp> {
       });
     });
 
-    _isAdvertisingSubscription =
-        beaconBroadcast.getAdvertisingStateChange().listen((isAdvertising) {
+    _isAdvertisingSubscription = beaconBroadcast.getAdvertisingStateChange().listen((isAdvertising) {
       setState(() {
         _isAdvertising = isAdvertising;
       });
@@ -57,32 +58,32 @@ class _MyAppState extends State<MyApp> {
               mainAxisSize: MainAxisSize.min,
               crossAxisAlignment: CrossAxisAlignment.start,
               children: <Widget>[
-                Text('Is transmission supported?',
-                    style: Theme.of(context).textTheme.headline5),
-                Text('$_isTransmissionSupported',
-                    style: Theme.of(context).textTheme.subtitle1),
+                Text('Is transmission supported?', style: Theme.of(context).textTheme.headline5),
+                Text('$_isTransmissionSupported', style: Theme.of(context).textTheme.subtitle1),
                 Container(height: 16.0),
-                Text('Is beacon started?', style: Theme.of(context).textTheme.headline5),
+                Text('Has beacon started?', style: Theme.of(context).textTheme.headline5),
                 Text('$_isAdvertising', style: Theme.of(context).textTheme.subtitle1),
                 Container(height: 16.0),
                 Center(
-                  child: RaisedButton(
+                  child: ElevatedButton(
                     onPressed: () {
                       beaconBroadcast
-                          .setUUID(UUID)
-                          .setMajorId(MAJOR_ID)
-                          .setMinorId(MINOR_ID)
-                          .setTransmissionPower(-59)
-                          .setIdentifier(IDENTIFIER)
-                          .setLayout(LAYOUT)
-                          .setManufacturerId(MANUFACTURER_ID)
+                          .setUUID(uuid)
+                          .setMajorId(majorId)
+                          .setMinorId(minorId)
+                          .setTransmissionPower(transmissionPower)
+                          .setAdvertiseMode(advertiseMode)
+                          .setIdentifier(identifier)
+                          .setLayout(layout)
+                          .setManufacturerId(manufacturerId)
+                          .setExtraData(extraData)
                           .start();
                     },
                     child: Text('START'),
                   ),
                 ),
                 Center(
-                  child: RaisedButton(
+                  child: ElevatedButton(
                     onPressed: () {
                       beaconBroadcast.stop();
                     },
@@ -90,13 +91,15 @@ class _MyAppState extends State<MyApp> {
                   ),
                 ),
                 Text('Beacon Data', style: Theme.of(context).textTheme.headline5),
-                Text('UUID: $UUID'),
-                Text('Major id: $MAJOR_ID'),
-                Text('Minor id: $MINOR_ID'),
-                Text('Tx Power: $TRANSMISSION_POWER'),
-                Text('Identifier: $IDENTIFIER'),
-                Text('Layout: $LAYOUT'),
-                Text('Manufacturer Id: $MANUFACTURER_ID'),
+                Text('UUID: $uuid'),
+                Text('Major id: $majorId'),
+                Text('Minor id: $minorId'),
+                Text('Tx Power: $transmissionPower'),
+                Text('Advertise Mode Value: $advertiseMode'),
+                Text('Identifier: $identifier'),
+                Text('Layout: $layout'),
+                Text('Manufacturer Id: $manufacturerId'),
+                Text('Extra data: $extraData'),
               ],
             ),
           ),
@@ -107,9 +110,7 @@ class _MyAppState extends State<MyApp> {
 
   @override
   void dispose() {
+    _isAdvertisingSubscription?.cancel();
     super.dispose();
-    if (_isAdvertisingSubscription != null) {
-      _isAdvertisingSubscription.cancel();
-    }
   }
 }