+++ lib/beacon_broadcast.dart
@@ -23,28 +23,25 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 
 class BeaconBroadcast {
-  static const String ALTBEACON_LAYOUT =
-      'm:2-3=beac,i:4-19,i:20-21,i:22-23,p:24-24,d:25-25';
-  static const String EDDYSTONE_TLM_LAYOUT =
-      'x,s:0-1=feaa,m:2-2=20,d:3-3,d:4-5,d:6-7,d:8-11,d:12-15';
-  static const String EDDYSTONE_UID_LAYOUT =
-      's:0-1=feaa,m:2-2=00,p:3-3:-41,i:4-13,i:14-19';
+  static const String ALTBEACON_LAYOUT = 'm:2-3=beac,i:4-19,i:20-21,i:22-23,p:24-24,d:25-25';
+  static const String EDDYSTONE_TLM_LAYOUT = 'x,s:0-1=feaa,m:2-2=20,d:3-3,d:4-5,d:6-7,d:8-11,d:12-15';
+  static const String EDDYSTONE_UID_LAYOUT = 's:0-1=feaa,m:2-2=00,p:3-3:-41,i:4-13,i:14-19';
   static const String EDDYSTONE_URL_LAYOUT = 's:0-1=feaa,m:2-2=10,p:3-3:-41,i:4-21v';
   static const String URI_BEACON_LAYOUT = 's:0-1=fed8,m:2-2=00,p:3-3:-41,i:4-21v';
 
-  String _uuid;
-  int _majorId;
-  int _minorId;
-  int _transmissionPower;
-  String _identifier = "";
-  String _layout;
-  int _manufacturerId;
+  String? _uuid;
+  int? _majorId;
+  int? _minorId;
+  int? _transmissionPower;
+  int? _advertiseMode;
+  String? _identifier;
+  String? _layout;
+  int? _manufacturerId;
+  List<int>? _extraData;
 
-  static const MethodChannel _methodChannel =
-      const MethodChannel('pl.pszklarska.beaconbroadcast/beacon_state');
+  static const MethodChannel _methodChannel = const MethodChannel('pl.pszklarska.beaconbroadcast/beacon_state');
 
-  static const EventChannel _eventChannel =
-      const EventChannel('pl.pszklarska.beaconbroadcast/beacon_events');
+  static const EventChannel _eventChannel = const EventChannel('pl.pszklarska.beaconbroadcast/beacon_events');
 
   /// Sets UUID for beacon.
   ///
@@ -101,6 +98,22 @@ class BeaconBroadcast {
     return this;
   }
 
+  /// Sets advertise mode for beacon.
+  ///
+  /// Advertise mode determines advertising frequency and power consumption.
+  ///
+  /// This parameter is **Android only** (it has no effect on iOS). It is optional, if not set, the default value will be [AdvertiseMode.balanced].
+  /// You can use one of the options:
+  /// <ul>
+  /// <li>[AdvertiseMode.lowPower] Consumes less energy, but larger broadcast interval
+  /// <li>[AdvertiseMode.balanced] default: Balance between energy usage and broadcast interval
+  /// <li>[AdvertiseMode.lowLatency] Consumes more energy, but smaller broadcast interval
+  /// </ul>
+  BeaconBroadcast setAdvertiseMode(AdvertiseMode advertiseMode) {
+    _advertiseMode = _advertiseModeToInt(advertiseMode);
+    return this;
+  }
+
   /// Sets beacon layout.
   ///
   /// This parameter is **Android only**. It's optional, the default is [ALTBEACON_LAYOUT].
@@ -130,11 +143,32 @@ class BeaconBroadcast {
     return this;
   }
 
+  /// Sets extra data.
+  ///
+  /// This parameter is **Android only**. If beacon layout allows it, you can
+  /// add extra bytes to the data transmitted by the beacon.
+  /// Value must be within a range 0-255.
+  ///
+  /// For more information check section
+  /// [Adding extra data](https://github.com/pszklarska/beacon_broadcast#adding-extra-data)
+  /// in the documentation.
+  ///
+  /// **For iOS**, beacon layout doesn't allow to transmit any extra data.
+  ///
+  /// This parameter is optional.
+  BeaconBroadcast setExtraData(List<int> extraData) {
+    if (extraData.any((value) => value < 0 || value > 255)) {
+      throw new IllegalArgumentException("Illegal arguments! Extra data values must be within a byte range 0-255");
+    }
+    _extraData = extraData;
+    return this;
+  }
+
   /// Starts beacon advertising.
   ///
   /// Before starting you must set  [_uuid].
   /// For the default layout, parameters [_majorId], [_minorId] are also required.
-  /// Other parameters as [_identifier], [_transmissionPower], [_layout], [_manufacturerId] are optional.
+  /// Other parameters as [_identifier], [_transmissionPower], [_advertiseMode], [_layout], [_manufacturerId] are optional.
   ///
   /// For Android, beacon layout is by default set to AltBeacon (check more details here: [AltBeacon - Transmitting as a Beacon](https://altbeacon.github.io/android-beacon-library/beacon-transmitter.html)).
   /// On Android system, it's required to have Bluetooth turn on and to give app permission to location.
@@ -144,15 +178,12 @@ class BeaconBroadcast {
   /// continue running in the foreground to broadcast the needed Bluetooth signals. If the user
   /// quits the app, the system stops advertising the device as a peripheral over Bluetooth.
   Future<void> start() async {
-    if (_uuid == null || _uuid.isEmpty) {
-      throw new IllegalArgumentException(
-          "Illegal arguments! UUID must not be null or empty: UUID: $_uuid");
+    if (_uuid == null || _uuid!.isEmpty) {
+      throw new IllegalArgumentException("Illegal arguments! UUID must not be null or empty: UUID: $_uuid");
     }
 
-    if ((_layout == null || _layout == ALTBEACON_LAYOUT) &&
-        (_majorId == null || _minorId == null)) {
-      throw new IllegalArgumentException(
-          "Illegal arguments! MajorId and minorId must not be null or empty: "
+    if ((_layout == null || _layout == ALTBEACON_LAYOUT) && (_majorId == null || _minorId == null)) {
+      throw new IllegalArgumentException("Illegal arguments! MajorId and minorId must not be null or empty: "
           "majorId: $_majorId, minorId: $_minorId");
     }
 
@@ -161,9 +192,11 @@ class BeaconBroadcast {
       "majorId": _majorId,
       "minorId": _minorId,
       "transmissionPower": _transmissionPower,
+      "advertiseMode": _advertiseMode,
       "identifier": _identifier,
       "layout": _layout,
       "manufacturerId": _manufacturerId,
+      "extraData": _extraData,
     };
 
     await _methodChannel.invokeMethod('start', params);
@@ -175,7 +208,7 @@ class BeaconBroadcast {
   }
 
   /// Returns `true` if beacon is advertising
-  Future<bool> isAdvertising() async {
+  Future<bool?> isAdvertising() async {
     return await _methodChannel.invokeMethod('isAdvertising');
   }
 
@@ -190,15 +223,14 @@ class BeaconBroadcast {
   /// Checks if device supports transmission. For iOS it returns always true.
   ///
   /// Possible values (for Android):
-  /// * [BeaconStatus.SUPPORTED] device supports transmission
-  /// * [BeaconStatus.NOT_SUPPORTED_MIN_SDK] Android system version on the device is lower than 21
-  /// * [BeaconStatus.NOT_SUPPORTED_BLE] BLE is not supported on this device
-  /// * [BeaconStatus.NOT_SUPPORTED_CANNOT_GET_ADVERTISER] device does not have a compatible chipset
+  /// * [BeaconStatus.supported] device supports transmission
+  /// * [BeaconStatus.notSupportedMinSdk] Android system version on the device is lower than 21
+  /// * [BeaconStatus.notSupportedBle] BLE is not supported on this device
+  /// * [BeaconStatus.notSupportedCannotGetAdvertiser] device does not have a compatible chipset
   /// or driver
   Future<BeaconStatus> checkTransmissionSupported() async {
-    var isTransmissionSupported =
-        await _methodChannel.invokeMethod('isTransmissionSupported');
-    return fromInt(isTransmissionSupported);
+    var isTransmissionSupported = await _methodChannel.invokeMethod('isTransmissionSupported');
+    return _beaconStatusFromInt(isTransmissionSupported);
   }
 }
 
@@ -214,27 +246,47 @@ class IllegalArgumentException implements Exception {
 
 enum BeaconStatus {
   /// Device supports transmitting as a beacon
-  SUPPORTED,
+  supported,
 
   /// Android system version on the device is too low (min. is 21)
-  NOT_SUPPORTED_MIN_SDK,
+  notSupportedMinSdk,
 
   /// Device doesn't support Bluetooth Low Energy
-  NOT_SUPPORTED_BLE,
+  notSupportedBle,
 
   /// Device's Bluetooth chipset or driver doesn't support transmitting
-  NOT_SUPPORTED_CANNOT_GET_ADVERTISER
+  notSupportedCannotGetAdvertiser
 }
 
-BeaconStatus fromInt(int value) {
-  switch (value) {
-    case 0:
-      return BeaconStatus.SUPPORTED;
-    case 1:
-      return BeaconStatus.NOT_SUPPORTED_MIN_SDK;
-    case 2:
-      return BeaconStatus.NOT_SUPPORTED_BLE;
-    default:
-      return BeaconStatus.NOT_SUPPORTED_CANNOT_GET_ADVERTISER;
-  }
+Map<int, BeaconStatus> _intToBeaconStatus = {
+  0: BeaconStatus.supported,
+  1: BeaconStatus.notSupportedMinSdk,
+  2: BeaconStatus.notSupportedBle,
+};
+
+BeaconStatus _beaconStatusFromInt(int? value) {
+  if (!_intToBeaconStatus.containsKey(value)) return BeaconStatus.notSupportedCannotGetAdvertiser;
+  return _intToBeaconStatus[value]!;
+}
+
+enum AdvertiseMode {
+  /// Consumes less energy, but larger broadcast interval
+  lowPower,
+
+  /// Balance between energy usage and broadcast interval
+  balanced,
+
+  /// Consumes more energy, but smaller broadcast interval
+  lowLatency,
+}
+
+Map<int, AdvertiseMode> _intToAdvertiseMode = {
+  0: AdvertiseMode.lowPower,
+  1: AdvertiseMode.balanced,
+  2: AdvertiseMode.lowLatency,
+};
+
+int? _advertiseModeToInt(AdvertiseMode advMode) {
+  if (!_intToAdvertiseMode.containsValue(advMode)) return null;
+  return _intToAdvertiseMode.entries.firstWhere((element) => element.value == advMode).key;
 }