+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.3.0
+
+Added support for null safety
+Added support for nullable identifiers (e.g. for the Eddystone layout)
+
+## 0.2.3
+
+Fixed data fields support on Android
+
+## 0.2.2
+
+Added support for setting data fields on Android
+Added support for setting advertisement mode on Android
+
 ## 0.2.1
 
 Updated the documentation